✨ Add final amount of water in the model

+++ TimerH2O/TimerH2O/model.swift
@@ -10,6 +10,6 @@ import Foundation
 
 struct Model {
     let idx: String
-    let water: Int
+    let water: Double
     let interval: TimeInterval
 }
